requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.5
django-allauth==0.57.0
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
